+++ packages/werkbank/lib/src/addons/src/constraints/src/_internal/ruler_overlay.dart
@@ -188,7 +188,8 @@ enum ConstraintsMode {
   // Those modes support using both axes at the same time
   min,
   max,
-  bothTight;
+  bothTight
+  ;
 
   bool get supportsBothAxes => this != ConstraintsMode.tightOneAxis;