+++ packages/werkbank/CHANGELOG.md
@@ -4,8 +4,21 @@
     - TODO: Mention replacement.
   - Rename PanelTabsController to Sections controller. TODO: even mention this?
 
+# 0.18.2
+- Add "Side by Side" semantics mode to semantics inspector.
+  - In this mode, the normal use case view is shown on one side and its semantics visualization on the other side.
+  - This allows you to interact with the use case as well as select semantics nodes at the same time.
+    In contrast the "Inspection" and "Overlay" modes allow you to only do one.
+  - You can switch between a vertical and horizontal split.
+
+# 0.18.1+1
+- Limit `slang` version to `<4.11.0` because of breaking change made without raising the major version.
+
 # 0.18.1
 - Add support for flutter 3.38.x
+
+# 0.18.0+1
+- Limit `slang` version to `<4.11.0` because of breaking change made without raising the major version.
 
 # 0.18.0
 - Add new required `themeBrightnessBuilder` parameter to `AppConfig()` and `AppConfig.widget()` constructors. **(BREAKING CHANGE)**